Refactor: Move ENVIRONMENT_CONFIG to a dedicated environment.config.ts and update imports
